requirements.txt
paramiko==2.7.1
pywin32==227
bypy==1.6.10
numpy==1.18.1
selenium==3.141.0
requests==2.18.4
opencv_python==4.4.0.44
Pillow==8.0.0
win32gui==221.6
